+++ public/components/health-check/lib/index.ts
@@ -1,3 +1,0 @@
-export { checkKibanaSettings } from './check-kibana-settings';
-export { checkKibanaSettingsTimeFilter } from './check-time-filter';
-export { checkKibanaSettingsMaxBuckets } from './check-max-buckets';