+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 - Removed module titles [#3160](https://github.com/wazuh/wazuh-kibana-app/pull/3160)
 - Removed Sha1 field from registry key detail [#3189](https://github.com/wazuh/wazuh-kibana-app/pull/3189)
+- Refactored the Health check component [#3197](https://github.com/wazuh/wazuh-kibana-app/pull/3197)
 
 ### Fixed